backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
torch==2.0.0
torchvision==0.15.1
pillow==9.4.0
numpy==1.24.2
scikit-image==0.19.3
python-dotenv==0.21.1
Werkzeug==2.2.3
gunicorn==20.1.0
opencv-python==4.7.0.72